use Input directly
