+++ Assets/Rector/Scripts/Editor/VfxInputBehaviourEditor.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace Rector.Editor
 {
-    [CustomEditor(typeof(VfxInputSlot))]
+    [CustomEditor(typeof(VfxInputSlotBehaviour))]
     public class VfxInputBehaviourEditor : UnityEditor.Editor
     {
         public override void OnInspectorGUI()
@@ -14,7 +14,7 @@ namespace Rector.Editor
             // FIXME: VFXGraphのエディタを開かないとうまく反映できない
             if (GUILayout.Button("Reset Properties"))
             {
-                var vfxView = (VfxInputSlot) target;
+                var vfxView = (VfxInputSlotBehaviour) target;
                 var props = VfxAssetReader.GetPropertyInputs(vfxView.VisualEffect);
                 vfxView.ResetProperties(props);
                 EditorUtility.SetDirty(vfxView);
@@ -23,7 +23,7 @@ namespace Rector.Editor
             // FIXME: VFXGraphのエディタを開かないとうまく反映できない
             if (GUILayout.Button("Reset Events"))
             {
-                var vfxView = (VfxInputSlot) target;
+                var vfxView = (VfxInputSlotBehaviour) target;
                 vfxView.ResetEvents(VfxAssetReader.GetEventNames(vfxView.VisualEffect));
                 EditorUtility.SetDirty(vfxView);
             }
